+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.1.10 - ????
+
+### Features
+
+-   Add advanced option `predict_single_probability_for_binary_classification`
+    to generate prediction tensors of shape [batch_size, 2] for binary
+    classification model.
+
 ## 0.1.9 - 2021-08-31
 
 ### Features